Provide access to `idxStr`
`idxStr` is a `char*` in the signature, but this is C and `char*` is
really just bad wording for `this might be anything, maybe some bytes`.
SQLite explicitly makes the meaning of this value defined by the module
implementation.
Some modules might make this a pointer to some array of integers (`int*`) or something even more wild.
We just don't know. The best we can do is pass on that pointer and let
the user handle it.
To do anything we need to be able to set it though.